♻️ refactor(progress): colocate progress hooks in hooks/ subfolder
Move useProgress, useProgressKeyDown, useWaveformMode, useWavesurfer
and the isLiveTrack helper into a Progress/hooks/ subfolder with an
index.ts barrel, mirroring Interface/hooks and Context/hooks.

Pure file move via git mv — no logic, signature, or type change.
Consumers import through the ./hooks barrel; test imports updated to
../hooks/*. Behavior identical: 419 tests pass, build succeeds.
