+++ package/src/components/AudioPlayer/Interface/Controller/Input/Progress/hooks/useWavesurfer.ts
